power added to velocity, projectile needed adding

+++ src/player/resources.rs
@@ -2,12 +2,12 @@ use bevy::prelude::*;
 
 #[derive(Resource, Debug)]
 pub struct Power {
-    pub total: u8,
+    pub total: f32,
     pub reverse: bool,
 }
 
 impl Default for Power {
     fn default() -> Power {
-        Power { total: 0, reverse: false}
+        Power { total: 0., reverse: false}
     }
 }
